requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
outcome==1.2.0
packaging==23.0
pydantic==1.10.7
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
starlette==0.26.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
webdriver-manager==3.8.5
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
